Update transcribe.py
Automatically selects the device on which the models should be loaded, either CUDA or CPU.
If Cuda is not available it will run on CPU. It also makes it easy to change models, the "medium" model is used by default. If you want to change models just replace the word "medium" with (tiny, base, small or large).